Add Export test on bluepirnts page to address PR#604
1. Add blueprint export test on blueprints page.
2. Move apiFetchTest() function to utils/commands.js because it
will be re-used by api.test.js and blueprints.test.js
